Move datafusion delta table factory util into tests/common
