+++ rust/tests/common/mod.rs
@@ -1,4 +1,5 @@
-#![deny(warnings)]
+#![allow(dead_code)]
+#![allow(unused_variables)]
 
 use bytes::Bytes;
 use deltalake::action::{self, Add, Remove};
@@ -15,6 +16,8 @@ use tempdir::TempDir;
 #[cfg(feature = "azure")]
 pub mod adls;
 pub mod clock;
+#[cfg(feature = "datafusion-ext")]
+pub mod datafusion;
 #[cfg(any(feature = "s3", feature = "s3-rustls"))]
 pub mod s3;
 pub mod schemas;
